wip: fix node id to support language
Previously "vid" was being used as the id of individual nodes,
but nid apparently is required to support the lanuages links.
This commit changes vid for nid, but also refactors the DB
creation since a fresh start is required. This commit is not
yet functional though because all nodes must be created before
starting link creation.